OnPoint-004 Move tabs menu to assignments activity [GM,MG]
By doing so we leave the main activity unchanged, so we are not
clashing with the other devs. The assignments activity has now the
tabbed menu on top. Each "content" of the tab (the body) is designed
in the fragments_*.xml. In the Assignments tab there is an
(empty) recyclerview with the list of tasks to be filled.
