+++ app/src/main/AndroidManifest.xml
@@ -9,11 +9,11 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.OnPoint">
-        <activity android:name=".MainActivity"
-            android:label="@string/title_activity_main_tabbed"
-            android:theme="@style/Theme.OnPoint.NoActionBar">
+        <activity android:name=".AssignmentsListActivity"></activity>
+        <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
+
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
