+++ app/src/main/java/at/tugraz/onpoint/AssignmentsListActivity.kt
@@ -2,10 +2,18 @@ package at.tugraz.onpoint
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import androidx.viewpager.widget.ViewPager
+import at.tugraz.onpoint.ui.main.SectionsPagerAdapter
+import com.google.android.material.tabs.TabLayout
 
 class AssignmentsListActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_assignments_list)
+        val sectionsPagerAdapter = SectionsPagerAdapter(this, supportFragmentManager)
+        val viewPager: ViewPager = findViewById(R.id.view_pager)
+        viewPager.adapter = sectionsPagerAdapter
+        val tabs: TabLayout = findViewById(R.id.tabs)
+        tabs.setupWithViewPager(viewPager)
     }
 }
