+++ app/src/main/res/values/strings.xml
@@ -1,7 +1,8 @@
 <resources>
     <string name="app_name">OnPoint</string>
-    <string name="title_activity_main_tabbed">MainTabbedActivity</string>
+    <string name="title_activity_main">MainActivity</string>
     <string name="tab_title_main">Main</string>
     <string name="tab_title_todo">Todo</string>
     <string name="tab_title_assignments">Assign.</string>
+    <string name="assignments">Assignments</string>
 </resources>
